+++ src/sys/dev/usb/uhci_pci.c
@@ -204,7 +204,8 @@ uhci_pci_attach(device_t self)
 
 	if (sc->sc_hw_ptr == NULL) {
 		device_printf(self, "Could not allocate DMA-able "
-			      "memory, %d bytes!\n", sizeof(*(sc->sc_hw_ptr)));
+			      "memory, %d bytes!\n", 
+			      (int32_t)sizeof(*(sc->sc_hw_ptr)));
 		return ENXIO;
 	}
